requirements.txt
django>=3.0,<4.0
psycopg2-binary>=2.8
celery
redis
django-celery-beat
django-celery-results
django-allauth
BeautifulSoup4
lxml
pandas
numpy
xlrd
finnhub-python
yfinance